Configuring the main window
We are first gonna create 3 lists, each of them displaying the artist name, their albums, and the songs of the respective album. When we click on an artist, the album list will show all of their albums, and when we click on an album, we get displayed all the songs in a third list, on that album. In order to get this response from our lists, the album and song list boxes will include a variable type that will help us change the albums or songs lists, based on the artist or albums we select.